Use vim.mpack by default
- The precedence is now:
  1. vim.mpack
  2. impatient.cachepack
  3. require('mpack')

- Put cachepack back into its own module.
- Restructure tests
